requirements.txt
pygsheets
pandas
python-dotenv
requests
discord.py
schedule
pytz